+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Four blocksparsity layouts from DeepSpeed [#320]
+- Support several initialization options [#312]
+
 
 ## [0.0.11] - 2022-05-30
 ### Fixed